% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allmax.R
\name{allmax}
\alias{allmax}
\title{Find all maxima}
\usage{
allmax(vector)
}
\arguments{
\item{vector}{A numerical vector}
}
\value{
A vector of the indices of the maxima
}
\description{
Find indices of all numbers equal to the maximum in a vector
}
\examples{
allmax(c(5,3,-7,5))
allmax(1:10)

}
