% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNash.2.R
\name{checkNash.2}
\alias{checkNash.2}
\title{Pure Nash checker (special)}
\usage{
checkNash.2(M, N, A, B)
}
\arguments{
\item{M}{Number of rows}

\item{N}{Number of columns}

\item{A}{Payoff matrix of Player A}

\item{B}{Payoff matrix of Player B}
}
\value{
A data frame of Nash Equilibria
}
\description{
Find all Nash equilibria (for use in the function \code{\link{nashGame}})
}
