% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNash.R
\name{checkNash}
\alias{checkNash}
\title{Pure Nash checker}
\usage{
checkNash(game)
}
\arguments{
\item{game}{An object of class "Game"}
}
\value{
A data frame of Nash Equilibria
}
\description{
Find all Nash equilibria in a game
}
\examples{
example_game <- nashGame(4,3,0,30)
nashEq <- checkNash(example_game)

}
