% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domGame.R
\name{domGame}
\alias{domGame}
\title{Dominant strategy game generator}
\usage{
domGame(M, N, min, max)
}
\arguments{
\item{M}{Number of strategies of Player A (rows)}

\item{N}{Number of strategies of Player B (columns)}

\item{min}{Smallest payoff}

\item{max}{Biggest payoff}
}
\value{
An object of class "Game"
}
\description{
Create a game with a unique equilibrium in dominant strategies
}
\examples{
example_game <- domGame(4,3,0,30)

}
