% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classGame.R
\docType{class}
\name{Game-class}
\alias{Game-class}
\title{New class for games}
\arguments{
\item{type}{A string to describe the game}

\item{M}{Number of strategies of Player A (rows)}

\item{N}{Number of strategies of Player B (columns)}

\item{payoff.A}{A numerical matrix of payoffs of Player A}

\item{payoff.B}{A numerical matrix of payoffs of Player B}

\item{sol.A}{A list of vectors of strategies of Player A}

\item{sol.B}{A list of vectors of strategies of Player B}

\item{sol.full}{A dataframe of game solutions, formatted according to game type}
}
\description{
New class for games
}
