% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iesdsGame.R
\name{iesdsGame}
\alias{iesdsGame}
\title{IESDS game generator}
\usage{
iesdsGame(M, N, min, max, maxdiff)
}
\arguments{
\item{M}{Number of strategies of Player A (rows)}

\item{N}{Number of strategies of Player B (columns)}

\item{min}{Approximate smallest payoff}

\item{max}{Approximate biggest payoff}

\item{maxdiff}{Maximum difference between a dominated strategy and the best strategy}
}
\value{
An object of class "Game"
}
\description{
Create a game with a unique equilibrium found by IESDS
}
\examples{
example_game <- iesdsGame(4,3,0,30,5)

}
