% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedGame.R
\name{mixedGame}
\alias{mixedGame}
\title{Mixed strategy game generator}
\usage{
mixedGame(M)
}
\arguments{
\item{M}{Number of strategies of Player A (rows) and Player B (columns)}
}
\value{
An object of class "Game"
}
\description{
Create a game with at least one mixed strategy equilibrium
}
\examples{
example_game <- mixedGame(4)

}
