% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedVec.R
\name{mixedVec}
\alias{mixedVec}
\title{Vector modification for mixedGame.R}
\usage{
mixedVec(payoffVec, strat, c)
}
\arguments{
\item{payoffVec}{A vector of payoffs, corresponding to a single pure strategy}

\item{strat}{The mixed strategy of the opponent}

\item{c}{The desired expected payoff}
}
\value{
A modified vector of payoffs
}
\description{
Given a vector of payoffs and the opponent's mixed strategy the
function modifies the payoffs so the expected payoff of the strategy equals
the given constant. Built for use in the function \code{\link{mixedGame}}.
}
\examples{
mixedVec(c(-5, 3, 7), c(0.2, 0.5, 0.3), 4)

}
