% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nashGame.R
\name{nashGame}
\alias{nashGame}
\title{Nash game generator}
\usage{
nashGame(M, N, min, max)
}
\arguments{
\item{M}{Number of strategies of Player A (rows)}

\item{N}{Number of strategies of Player B (columns)}

\item{min}{Smallest payoff}

\item{max}{Biggest payoff}
}
\value{
An object of class "Game"
}
\description{
Create a game with one or more pure strategy Nash equilibria
}
\examples{
example_game <- nashGame(4,3,0,30)

}
