% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalGame.R
\name{normalGame}
\alias{normalGame}
\title{Normal form representation}
\usage{
normalGame(game)
}
\arguments{
\item{game}{An object of class "Game"}
}
\value{
A matrix with a normal form representation.
}
\description{
Creates a normal form representation viewable as an R matrix.
To make the game easier to read, each row and column of the original
payoff matrices are duplicated. The payoffs of A are then in the bottom left
of each 2x2 cell, and the payoffs of B are in the top right.
}
\examples{
example_game <- nashGame(4,3,0,30)
normal_form <- normalGame(example_game)

}
