% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalGame2.R
\name{normalGame2}
\alias{normalGame2}
\title{Normal form representation (for Shiny)}
\usage{
normalGame2(game)
}
\arguments{
\item{game}{An object of class "Game"}
}
\value{
A matrix with a normal form representation.
}
\description{
Creates a normal form representation viewable as an R matrix.
To ensure compatibility with Shiny tables, each outcome is represented
by a single cell, with the payoffs of A and B separated by a comma.
}
\examples{
example_game <- nashGame(4,3,0,30)
normal_form <- normalGame2(example_game)

}
