% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalTexGame.R
\name{normalTexGame}
\alias{normalTexGame}
\title{TeX output of a game}
\usage{
normalTexGame(game, showOutput = TRUE, showNash = FALSE, showSol = FALSE,
  fileOut = FALSE)
}
\arguments{
\item{game}{An object of class "Game"}

\item{showOutput}{A boolean to output the text to console. Overridden by \code{fileOut}. Defaults to \code{TRUE}.}

\item{showNash}{A boolean to switch on underlining of pure strategy Nash equilibria. Defaults to \code{FALSE}.}

\item{showSol}{A boolean to output the solution text. Defaults to \code{FALSE}.}

\item{fileOut}{A boolean to write directly to a TeX file. Defaults to \code{FALSE}.}
}
\value{
A character vector with each cell being one line of the TeX code which can be outputted
with \code{writeLines()}. By default, the output is printed directly to the console.
}
\description{
Creates a normal form representation of a game and exports it
in LaTeX format. Can also display pure strategy Nash equilibria
as well as solutions adapted to each type of game. By default, exports
to console, but can also write to a TeX file.
}
\examples{
example_game <- nashGame(4,3,0,30)
normalTexGame(example_game, showNash = TRUE)
normalTexGame(example_game, showSol = TRUE, fileOut = TRUE)

}
