% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutionHTML.R
\name{solutionHTML}
\alias{solutionHTML}
\title{HTML output of a game solution}
\usage{
solutionHTML(game)
}
\arguments{
\item{game}{An object of class "Game"}
}
\description{
Writes the solution of the game according to game type with html
formatting. Primarily intended for use in shiny.
}
\examples{
example_game <- nashGame(4,3,0,30)
solutionHTML(example_game)

}
