% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeXrow.A.R
\name{TeXrow.A}
\alias{TeXrow.A}
\title{TeX row (A)}
\usage{
TeXrow.A(rowA)
}
\arguments{
\item{rowA}{A row of payoffs of Player A}
}
\value{
A string of TeX output
}
\description{
Generates a row of TeX code for the payoffs of Player A. For use in \code{\link{normalTexGame}}.
}
