% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeXrow.B.R
\name{TeXrow.B}
\alias{TeXrow.B}
\title{TeX row (B)}
\usage{
TeXrow.B(rowB)
}
\arguments{
\item{rowB}{A row of payoffs of Player B}
}
\value{
A string of TeX output
}
\description{
Generates a row of TeX code for the payoffs of Player B. For use in \code{\link{normalTexGame}}.
}
